Relocate the events
