+++ src/main/java/com/wisdom/quote/eventsourcing/events/BaseQuoteEvent.java
@@ -1,4 +1,4 @@
-package com.wisdom.quote.writemodel.event;
+package com.wisdom.quote.eventsourcing.events;
 
 import com.wisdom.eventstoredb.utils.Event;
